muda modelos do banco para a pasta database e corrige erros nos nomes
